fix vps_get_password stub namespace and make cpuCores an int
Two separate root causes behind the SetVMAdminPassword error and the cpuCores
formula failure:

- tests/PluginTest.php declares `namespace Detain\MyAdminHyperv\Tests;` in
  statement form, so the inline `function vps_get_password()` it defined landed
  in the Tests namespace. Plugin.php calls it unqualified from
  Detain\MyAdminHyperv, which resolves to that namespace then global -- never
  Tests -- so the call died with "undefined function
  Detain\MyAdminHyperv\vps_get_password()". Moved it to tests/stubs.php inside a
  `namespace Detain\MyAdminHyperv { }` block (same pattern as the whmsonic
  package) and pointed phpunit at a tests/bootstrap.php that loads stubs before
  the autoloader.
- getSoapCallParams('UpdateVM') built cpuCores with ceil(), which returns a
  float, so the SOAP payload carried 1.0 where every sibling integer field
  (ramMB, minimumOps, maximumOps, vhdSize) carried an int. Cast to int at both
  sites that build the UpdateVM parameters.

The stub now encodes both of its arguments in the return value, so the test
asserts newPassword === 'generated-pass-42-100' and proves getSoapCallParams
passes vps_id and vps_custid through in the right order; also tightened
existingPassword from assertArrayHasKey to an equality check.

+++ src/Plugin.php
@@ -179,7 +179,9 @@ class Plugin
         } elseif ($call == 'UpdateVM') {
             return [
                 'vmId' => $serviceInfo['vps_vzid'],
-                'cpuCores' => ceil((($serviceInfo['vps_slices'] - 2) / 2) + 1),
+                // ceil() returns a float; cpuCores is a whole-number core count,
+                // so cast it like the sibling integer fields (ramMB, minimumOps).
+                'cpuCores' => (int) ceil((($serviceInfo['vps_slices'] - 2) / 2) + 1),
                 'ramMB' => 1 * VPS_SLICE_RAM * $serviceInfo['vps_slices'],
                 'bootFromCD' => false,
                 'numLockEnabled' => true,
@@ -475,7 +477,8 @@ class Plugin
         $extra['response'] = $response;
         $update_parameters = [
             'vmId' => $serviceInfo['vzid'],
-            'cpuCores' => ceil((($serviceInfo['vps_slices'] - 2) / 2) + 1),
+            // ceil() returns a float; cpuCores is a whole-number core count.
+            'cpuCores' => (int) ceil((($serviceInfo['vps_slices'] - 2) / 2) + 1),
             'ramMB' => 1 * VPS_SLICE_RAM * $serviceInfo['vps_slices'],
             'bootFromCD' => false,
             'numLockEnabled' => true,
